refactor(ui): extract split button from base button
Move split layout, controlled menu state, segment sizing, and trigger accessibility into a focused SplitButton component. Share the ordinary button presentation primitives, migrate every live caller, and keep the base Button API limited to single-action controls.
