+++ src/components/Button/index.test.ts
@@ -2,7 +2,7 @@ import { readFileSync } from "node:fs";
 import { resolve } from "node:path";
 import React from "react";
 import { renderToStaticMarkup } from "react-dom/server";
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 
 import Button from ".";
 
@@ -35,70 +35,13 @@ function contrastRatio(first: string, second: string): number {
   );
 }
 
-function renderSplitDropdownClassName(
-  variant:
-    | "primary"
-    | "secondary"
-    | "danger"
-    | "warning"
-    | "success"
-    | "merged",
-  dropdownVisible = false
-): string {
-  const markup = renderToStaticMarkup(
-    React.createElement(
-      Button,
-      {
-        variant,
-        dropdownMenu: React.createElement("div"),
-        onDropdownClick: vi.fn(),
-        dropdownVisible,
-      },
-      "Action"
-    )
-  );
-  const buttonClassNames = [...markup.matchAll(/<button[^>]*class="([^"]*)"/g)];
-  return buttonClassNames.at(-1)?.[1] ?? "";
-}
-
-describe("Button split dropdown segment", () => {
-  it("uses the success tone while hovered or open", () => {
-    expect(renderSplitDropdownClassName("success")).toContain(
-      "enabled:hover:bg-success-5"
-    );
-    expect(renderSplitDropdownClassName("success", true)).toContain(
-      "bg-success-5 enabled:hover:bg-success-5"
-    );
-  });
-
-  it("uses GitHub purple for the merged variant and its split state", () => {
-    expect(renderSplitDropdownClassName("merged")).toContain(
-      "enabled:hover:bg-merged-hover"
-    );
-    expect(renderSplitDropdownClassName("merged", true)).toContain(
-      "bg-merged-hover enabled:hover:bg-merged-hover"
-    );
+describe("Button", () => {
+  it("uses GitHub purple for the merged variant", () => {
     const markup = renderToStaticMarkup(
       React.createElement(Button, { variant: "merged" }, "Merged")
     );
     expect(markup).toContain("bg-merged");
     expect(markup).toContain("text-merged-contrast");
-  });
-
-  it.each([
-    ["primary", "primary"],
-    ["danger", "danger"],
-    ["warning", "warning"],
-  ] as const)("uses the %s tone for its semantic variant", (variant, tone) => {
-    expect(renderSplitDropdownClassName(variant)).toContain(
-      `enabled:hover:bg-${tone}-5`
-    );
-  });
-
-  it("keeps neutral solid split buttons neutral", () => {
-    expect(renderSplitDropdownClassName("secondary")).toContain(
-      "enabled:hover:bg-fill-3"
-    );
   });
 
   it.each(["orgii_main.css", "orgii_dark.css", "orgii_high_contrast.css"])(
