+++ src/modules/shared/dataSource/RuntimeScanningPanelColumns.tsx
@@ -20,6 +20,7 @@ import {
   SETTINGS_TABLE_COL,
   type SettingsTableColumn,
 } from "@src/components/SettingsTable";
+import SplitButton from "@src/components/SplitButton";
 import Switch from "@src/components/Switch";
 import Tag from "@src/components/Tag";
 import {
@@ -179,24 +180,26 @@ export function buildRuntimeScanningPanelColumns({
                 // Importable sources have a cache, so offer two rescan modes via
                 // a split button: the main click runs Update (incremental
                 // re-sync); the caret opens Update / Clear + rescan (full rebuild).
-                <Button
+                <SplitButton
                   variant="secondary"
                   size="small"
                   iconOnly
-                  splitDropdownWidth={22}
+                  menuSegmentWidth={22}
                   loading={row.rescanning}
                   loadingSpinIcon
                   icon={<RefreshCw size={14} />}
+                  aria-label={t("rescan")}
                   title={t("rescan")}
                   onClick={() => void handleRescan(row, false)}
-                  dropdownVisible={openRescanMenu === row.probe.sourceId}
-                  onDropdownClick={(event) => {
+                  menuOpen={openRescanMenu === row.probe.sourceId}
+                  menuButtonLabel={t("rescan")}
+                  onMenuButtonClick={(event) => {
                     event.stopPropagation();
                     setOpenRescanMenu((current) =>
                       current === row.probe.sourceId ? null : row.probe.sourceId
                     );
                   }}
-                  dropdownMenu={
+                  menu={
                     <Dropdown
                       trigger="click"
                       position="bottom-end"
